Restructure to use userspace
