+++ keyboards/planck/keymaps/prog_qgmlwb/keymap.c
@@ -1,4 +1,4 @@
 // Copyright 2022 David Kristoffersen (@davidkristoffersen)
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#include "src/base.h"
+#include "keymap.h"
