+++ keyboards/planck/keymaps/prog_qgmlwb/src/config.h
@@ -2,6 +2,3 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #pragma once
-
-#define SWAP_GUI_ALT false
-#define SPLIT_SPACE false
