+++ users/davidkristoffersen/davidkristoffersen.c
@@ -1,10 +1,4 @@
 // Copyright 2022 David Kristoffersen (@davidkristoffersen)
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#pragma once
-
-#include QMK_KEYBOARD_H
-
-#include "config.h"
-
-#include "layouts.h"
+#include "davidkristoffersen.h"
