+++ users/davidkristoffersen/davidkristoffersen.h
@@ -1,0 +1,15 @@
+// Copyright 2022 David Kristoffersen (@davidkristoffersen)
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#pragma once
+
+#include "config.h"
+
+#include "norwegian.h"
+
+#include "util/functions.h"
+
+#include "hardware/left_space.h"
+#include "hardware/swap_gui_alt.h"
+
+#include "macros/macros.h"
