+++ users/davidkristoffersen/macros/dynamic.c
