+++ users/davidkristoffersen/macros/dynamic.h
@@ -3,6 +3,6 @@
 
 #pragma once
 
-#include "../base.h"
+#include "macros.h"
 
 bool process_dynamic(uint16_t keycode, keyrecord_t* record);
