+++ users/davidkristoffersen/macros/static.c
