+++ users/davidkristoffersen/rules.mk
@@ -1,0 +1,5 @@
+SRC += $(USER_PATH)/davidkristoffersen.c \
+		$(USER_PATH)/util/functions.c \
+		$(USER_PATH)/macros/macros.c \
+		$(USER_PATH)/macros/static.c \
+		$(USER_PATH)/macros/dynamic.c
