+++ users/davidkristoffersen/util/functions.c
